SNN/requirements.txt
torch
torchvision
pillow
matplotlib
numpy
scikit-learn
tqdm